requirements.txt
requests==2.31.0
schedule==1.2.1
PyYAML==6.0.1
snowflake-connector-python==3.8.0